Changelog: 1. Looped audience animations in Museum. 2. Added audience sound, curator speech and art gallery music. 3. Change to enlarge.cs script. 4. Updated README.md
