+++ Assets/Audio/Curator Speech 2.wav.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 0900b03abde0d9740b5e9a197b0d3b12
+timeCreated: 1459013213
+licenseType: Free
+AudioImporter:
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
