+++ Assets/Audio/Curator speech.wav.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 9870a2e6900f49542ae6c6184ceaf3a4
+timeCreated: 1459015016
+licenseType: Free
+AudioImporter:
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
