+++ Assets/Audio/crowd_milling_about.wav.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 64bb22ade61a8c44ba2e2b8e37f8d4e7
+timeCreated: 1459010237
+licenseType: Free
+AudioImporter:
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
