+++ Assets/Audio/music for art gallery.wav.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: b5aaa749e62b1464e89a4a260d2cfb76
+timeCreated: 1459015773
+licenseType: Free
+AudioImporter:
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
