+++ Assets/Scenes/Museum.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 68235eb3df9ccef4899d2675573adfb6
+folderAsset: yes
+timeCreated: 1459011987
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
