+++ Assets/Scenes/Museum/LightingData.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a7d18c7b1f5f0c142a3309d086805c42
+timeCreated: 1459012328
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
