+++ Assets/Scripts/CuratorAudio.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: d5779647d52d51140bef97e1763f98ba
+timeCreated: 1459013231
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
