+++ Assets/Scripts/panz.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: cdf460a144adb4c2388b7a56921a1d49
+timeCreated: 1458929482
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
