--- Assets/movebrush.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 27ffdd41944d443fca79c384364e5b1a
-timeCreated: 1458927633
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
