--- Assets/ray.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 81de78e32900b4d9c8c22656738256e4
-timeCreated: 1458929892
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
